Airflow Pipeline/requirements.txt
pyspark==3.1.2
pandas
psycopg2-binary==2.9.9
dotenv
sqlalchemy